requirements.txt
aiogram>=3.3.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
cryptography>=41.0.0
aiohttp>=3.9.0